feat(verifier): add evaluator backend facade
